+++ new-engine/src/client_handler.h
@@ -1,17 +1,20 @@
-#include<stdbool.h>
-
 #ifndef D_CLIENT_HANDLER
 #define D_CLIENT_HANDLER
 
+#include <stdbool.h>
+#include <stdlib.h>
+
 void Client_init(const char *server_ip);
+
 void Client_tick();
+
 void Client_quit();
 
 
 // Client player index, provided by the server.
 //  This is the "id" that the client and server use
 //  to control the proper player
-int Client_player_index;
+size_t Client_player_index;
 
 bool Client_got_new_inputs; // Did we get new inputs?
